Remove some deprecations to fix most of Behat scenarios

+++ src/Sylius/Behat/Context/Setup/PromotionContext.php
@@ -356,9 +356,9 @@ final class PromotionContext implements Context
      */
     public function thisPromotionGivesPercentageDiscountOnEveryProductInTheChannelAndInTheChannel(
         PromotionInterface $promotion,
-        int $firstPercentage,
+        float $firstPercentage,
         ChannelInterface $firstChannel,
-        int $secondPercentage,
+        float $secondPercentage,
         ChannelInterface $secondChannel,
     ): void {
         /** @var PromotionActionInterface $action */
